pyproject.toml
[project]
name = "hw1-asr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "datasets>=4.1.1",
    "editdistance>=0.8.1",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "ipython>=9.5.0",
    "isort>=6.0.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.6",
    "numpy==1.26.4",
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
    "pyqt6>=6.9.1",
    "soundfile>=0.13.1",
    "torch==2.2.0",
    "torch-audiomentations>=0.12.0",
    "torchaudio==2.2.0",
    "torchvision==0.17.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.0",
]
